Crate base_custom

base_custom

allows you to use any set of characters as your own numeric base and convert to and from decimal. This can be taken advantage of in various ways:

Mathematics: number conversion

Brute force sequencing

Rolling ciphers

Moderate information concealment

Other potential uses such as deriving music or art from numbers

To Include It

Add base_custom to your dependencies section of your Cargo.toml file.

[dependencies]
base_custom = "^0.1"

In your rust files where you plan to use it put this at the top

extern crate base_custom;
use base_custom::BaseCustom;

This is licensed under MIT or APACHE 2.0 at your option.

Structs

BaseCustom
The BaseCustom struct holds the information to perform number conversions via the gen and decimal methods.